pretty sure i got the getrunlistlayout working... need to move the code to get mft directory content into a GetDataAttributeLayout() function

+++ common.h
@@ -14,7 +14,7 @@ void ReadContent(std::ifstream* rawcontent, int8_t* tmpbuf, uint64_t offset, uin
 void ReturnUint32(uint32_t* tmp32, uint8_t* tmp8);
 void ReturnUint16(uint16_t* tmp16, uint8_t* tmp8);
 void ReturnUint64(uint64_t* tmp64, uint8_t* tmp8);
-void ReturnUint(unsigned int* tmp, uint8_t* tmp8, unsigned int length);
-void ReturnInt(int* tmp, int8_t* tmp8, unsigned int length);
+void ReturnUint(uint64_t* tmp, uint8_t* tmp8, unsigned int length);
+void ReturnInt(int64_t* tmp, int8_t* tmp8, unsigned int length);
 
 #endif
